Fix build errors and build session discovery APKs
Consumer App:
- Removed duplicate SessionInfo class
- Simplified main.dart (removed unused HomeScreen)
- Fixed provider location typo
- APK: 22.6MB

Provider App:
- Fixed Int64 to int conversion for timestamps
- APK: 22.4MB

Both APKs built successfully with full session discovery flow:
- Consumer: Lists providers, requests connection, waits for approval
- Provider: Creates session, receives requests, approves/denies

System 100% complete and ready for testing!

Commit 36

+++ provider_app/lib/services/request_watcher_service.dart
@@ -35,7 +35,7 @@ class RequestWatcherService {
             consumerId: notification.consumerId,
             consumerName: notification.consumerName,
             requestedAt: DateTime.fromMillisecondsSinceEpoch(
-              notification.requestedAt * 1000,
+              notification.requestedAt.toInt() * 1000,
             ),
           ));
         },
